requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0
PyYAML>=6.0
rapidfuzz>=3.6
aiosqlite>=0.20
